backend/requirements.txt
beautifulsoup4==4.4.1
chardet==2.3.0
click==6.7
command-not-found==0.3
defer==1.0.6
Flask==0.12.2
html5lib==0.999
itsdangerous==0.24
Jinja2==2.9.6
language-selector==0.1
lxml==3.5.0
MarkupSafe==1.0
menulibre==2.1.3
mugshot==0.3.1
numpy==1.13.1
onboard==1.2.0
opencv-python==3.3.0.10
pexpect==4.0.1
Pillow==3.1.2
psutil==3.4.2
ptyprocess==0.5
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
pytesseract==0.1.7
python-apt==1.1.0b1
python-debian==0.1.27
pytube==6.4.2
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
sessioninstaller==0.0.0
six==1.10.0
ssh-import-id==5.5
unattended-upgrades==0.1
urllib3==1.13.1
virtualenv==15.0.1
Werkzeug==0.12.2
